Move request/response data classes into new modules package so we can ignore the generated files from code coverage; add multi-module coverage reporter plugin
