+++ build.gradle
@@ -27,6 +27,20 @@ plugins {
     id("io.github.gradle-nexus.publish-plugin") version "1.1.0"
     id "org.jlleitschuh.gradle.ktlint" version "11.0.0"
     id "io.gitlab.arturbosch.detekt" version "1.22.0-RC3"
+    id 'nl.neotech.plugin.rootcoverage' version '1.5.3'
+}
+
+rootCoverage {
+    buildVariant "debug"
+    excludes.addAll(
+            "**/models/*",
+    )
+    generateHtml true
+    generateXml true
+    executeAndroidTests false
+    executeUnitTests true
+    includeUnitTestResults true
+    includeNoLocationClasses true
 }
 
 apply from: "${rootDir}/scripts/publish-root.gradle"
