+++ sdk/src/main/java/com/stytch/sdk/b2b/sessions/B2BSessionStorage.kt
@@ -1,7 +1,7 @@
 package com.stytch.sdk.b2b.sessions
 
-import com.stytch.sdk.b2b.network.B2BSessionData
-import com.stytch.sdk.b2b.network.MemberData
+import com.stytch.sdk.b2b.network.models.B2BSessionData
+import com.stytch.sdk.b2b.network.models.MemberData
 import com.stytch.sdk.common.Constants
 import com.stytch.sdk.common.StorageHelper
