+++ sdk/src/main/java/com/stytch/sdk/common/StytchExceptions.kt
@@ -1,6 +1,6 @@
 package com.stytch.sdk.common
 
-import com.stytch.sdk.common.network.StytchErrorResponse
+import com.stytch.sdk.common.network.models.StytchErrorResponse
 
 /**
  * Exception wrapper meant for use with Stytch-specific exceptions and provides a straightforward way to encompass any possible errors that come from the SDK.
