+++ sdk/src/main/java/com/stytch/sdk/common/Typealiases.kt
@@ -1,7 +1,7 @@
 // ktlint-disable filename
 package com.stytch.sdk.common
 
-import com.stytch.sdk.common.network.BasicData
+import com.stytch.sdk.common.network.models.BasicData
 
 /**
  * Type alias for StytchResult<BasicData> used for basic responses
