+++ sdk/src/main/java/com/stytch/sdk/common/network/models/CommonRequests.kt
@@ -1,4 +1,4 @@
-package com.stytch.sdk.common.network
+package com.stytch.sdk.common.network.models
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
