+++ sdk/src/main/java/com/stytch/sdk/consumer/Typealiases.kt
@@ -1,14 +1,14 @@
 package com.stytch.sdk.consumer
 
 import com.stytch.sdk.common.StytchResult
-import com.stytch.sdk.common.network.LoginOrCreateOTPData
-import com.stytch.sdk.common.network.StrengthCheckResponse
-import com.stytch.sdk.consumer.network.BiometricsAuthData
-import com.stytch.sdk.consumer.network.CreateResponse
-import com.stytch.sdk.consumer.network.DeleteAuthenticationFactorData
-import com.stytch.sdk.consumer.network.IAuthData
-import com.stytch.sdk.consumer.network.OAuthData
-import com.stytch.sdk.consumer.network.UserData
+import com.stytch.sdk.common.network.models.LoginOrCreateOTPData
+import com.stytch.sdk.common.network.models.StrengthCheckResponse
+import com.stytch.sdk.consumer.network.models.BiometricsAuthData
+import com.stytch.sdk.consumer.network.models.CreateResponse
+import com.stytch.sdk.consumer.network.models.DeleteAuthenticationFactorData
+import com.stytch.sdk.consumer.network.models.IAuthData
+import com.stytch.sdk.consumer.network.models.OAuthData
+import com.stytch.sdk.consumer.network.models.UserData
 
 /**
  * Type alias for StytchResult<AuthData> used for authentication responses
