+++ sdk/src/main/java/com/stytch/sdk/consumer/network/StytchApiService.kt
@@ -1,8 +1,10 @@
 package com.stytch.sdk.consumer.network
 
 import com.stytch.sdk.common.network.ApiService
-import com.stytch.sdk.common.network.CommonRequests
-import com.stytch.sdk.common.network.CommonResponses
+import com.stytch.sdk.common.network.models.CommonRequests
+import com.stytch.sdk.common.network.models.CommonResponses
+import com.stytch.sdk.consumer.network.models.ConsumerRequests
+import com.stytch.sdk.consumer.network.models.ConsumerResponses
 import retrofit2.http.Body
 import retrofit2.http.DELETE
 import retrofit2.http.GET
