+++ sdk/src/main/java/com/stytch/sdk/consumer/userManagement/UserManagementImpl.kt
@@ -5,7 +5,7 @@ import com.stytch.sdk.common.StytchResult
 import com.stytch.sdk.consumer.DeleteFactorResponse
 import com.stytch.sdk.consumer.UserResponse
 import com.stytch.sdk.consumer.network.StytchApi
-import com.stytch.sdk.consumer.network.UserData
+import com.stytch.sdk.consumer.network.models.UserData
 import com.stytch.sdk.consumer.sessions.ConsumerSessionStorage
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
