+++ sdk/src/test/java/com/stytch/sdk/b2b/network/StytchB2BApiServiceTest.kt
@@ -1,7 +1,8 @@
 package com.stytch.sdk.b2b.network
 
+import com.stytch.sdk.b2b.network.models.B2BRequests
 import com.stytch.sdk.common.network.ApiService
-import com.stytch.sdk.common.network.CommonRequests
+import com.stytch.sdk.common.network.models.CommonRequests
 import com.stytch.sdk.utils.verifyGet
 import com.stytch.sdk.utils.verifyPost
 import kotlinx.coroutines.runBlocking
