+++ sdk/src/test/java/com/stytch/sdk/b2b/organization/OrganizationImplTest.kt
@@ -1,9 +1,8 @@
 package com.stytch.sdk.b2b.organization
 
 import com.stytch.sdk.b2b.OrganizationResponse
-import com.stytch.sdk.b2b.network.Organization
-import com.stytch.sdk.b2b.network.OrganizationResponseData
 import com.stytch.sdk.b2b.network.StytchB2BApi
+import com.stytch.sdk.b2b.network.models.OrganizationResponseData
 import com.stytch.sdk.common.StytchDispatchers
 import com.stytch.sdk.common.StytchResult
 import io.mockk.MockKAnnotations
