+++ sdk/src/test/java/com/stytch/sdk/b2b/sessions/B2BSessionsImplTest.kt
@@ -2,8 +2,8 @@ package com.stytch.sdk.b2b.sessions
 
 import com.stytch.sdk.b2b.AuthResponse
 import com.stytch.sdk.b2b.extensions.launchSessionUpdater
-import com.stytch.sdk.b2b.network.IB2BAuthData
 import com.stytch.sdk.b2b.network.StytchB2BApi
+import com.stytch.sdk.b2b.network.models.IB2BAuthData
 import com.stytch.sdk.common.BaseResponse
 import com.stytch.sdk.common.EncryptionManager
 import com.stytch.sdk.common.StytchDispatchers
