+++ sdk/src/test/java/com/stytch/sdk/consumer/oauth/GoogleOneTapImplTest.kt
@@ -18,8 +18,8 @@ import com.stytch.sdk.common.oauth.GoogleOneTapProvider
 import com.stytch.sdk.common.sessions.SessionAutoUpdater
 import com.stytch.sdk.consumer.AuthResponse
 import com.stytch.sdk.consumer.extensions.launchSessionUpdater
-import com.stytch.sdk.consumer.network.AuthData
 import com.stytch.sdk.consumer.network.StytchApi
+import com.stytch.sdk.consumer.network.models.AuthData
 import com.stytch.sdk.consumer.sessions.ConsumerSessionStorage
 import io.mockk.MockKAnnotations
 import io.mockk.clearAllMocks
