+++ sdk/src/test/java/com/stytch/sdk/consumer/sessions/SessionStorageTest.kt
@@ -3,7 +3,7 @@ package com.stytch.sdk.consumer.sessions
 import com.stytch.sdk.common.Constants.PREFERENCES_NAME_SESSION_JWT
 import com.stytch.sdk.common.Constants.PREFERENCES_NAME_SESSION_TOKEN
 import com.stytch.sdk.common.StorageHelper
-import com.stytch.sdk.consumer.network.SessionData
+import com.stytch.sdk.consumer.network.models.SessionData
 import io.mockk.MockKAnnotations
 import io.mockk.clearAllMocks
 import io.mockk.every
